refactor: extract editor props to transformed props to its own hook
