+++ packages/core/components/Puck/components/Preview/components/edit-page.tsx
@@ -3,22 +3,16 @@ import { useShallow } from "zustand/react/shallow";
 
 import { useAppStore } from "../../../../../store";
 import { toComponent } from "../../../../../lib/data/to-component";
-import { getSlotTransform } from "../../../../../lib/field-transforms/default-transforms/slot-transform";
-import { useFieldTransformsTracked } from "../../../../../lib/field-transforms/use-field-transforms-tracked";
 import { expandNode } from "../../../../../lib/data/flatten-node";
 import { rootDroppableId } from "../../../../../lib/root-droppable-id";
 import {
   ComponentData,
   DefaultRootRenderProps,
-  FieldTransforms,
   RootData,
 } from "../../../../../types";
 
-import { DropZoneEditPure, DropZonePure } from "../../../../DropZone";
-import { getInlineTextTransform } from "../../../../../lib/field-transforms/default-transforms/inline-text-transform";
-import { getRichTextTransform } from "../../../../../lib/field-transforms/default-transforms/rich-text-transform";
-
-const slotFieldTransforms = getSlotTransform(DropZoneEditPure);
+import { DropZonePure } from "../../../../DropZone";
+import useEditorProps from "../../../../../lib/field-transforms/use-editor-props";
 
 /**
  * Renders the puck data as an editable page.
@@ -30,27 +24,6 @@ const EditPage = () => {
   );
   const config = useAppStore((s) => s.config);
   const metadata = useAppStore((s) => s.metadata);
-  const userFieldTransforms = useAppStore((s) => s.fieldTransforms);
-  const plugins = useAppStore((s) => s.plugins);
-
-  const combinedFieldTransforms = useMemo(
-    () => ({
-      ...slotFieldTransforms,
-      ...getInlineTextTransform(),
-      ...getRichTextTransform(),
-      ...plugins.reduce<FieldTransforms>((acc, plugin) => {
-        if (!plugin.fieldTransforms) return acc;
-
-        Object.keys(plugin.fieldTransforms).forEach((key) => {
-          acc[key as keyof typeof acc] = plugin.fieldTransforms![key];
-        });
-
-        return acc;
-      }, {}),
-      ...userFieldTransforms,
-    }),
-    [plugins, userFieldTransforms]
-  );
 
   // Root as object with slots still defaulted to null
   const rootAsComponent = useMemo(() => {
@@ -63,11 +36,7 @@ const EditPage = () => {
 
   // Get the props with stable slot render functions
   // (tracked only sees `null` for slot props across calls, so prev always === next)
-  const propsWithSlots = useFieldTransformsTracked(
-    config,
-    rootAsComponent,
-    combinedFieldTransforms
-  );
+  const propsWithSlots = useEditorProps({ component: rootAsComponent });
 
   // Build the final props to be passed to the user provided root render
   const renderProps: DefaultRootRenderProps = useMemo(() => {
